fix for 4.2 and 4.3.
The Method.invoke(Object obj, Object ...args) apparently doesn't properly handle a singleton byte[] array as an argument. Wrap that argument in an Object[] array and pass that in to work around the Varargs interpretation ambiguity.
The stream only needs to be synchronized against itself rather than on the global mutex. I don't think this increases performance given the code paths, but it is technically more correct.