fix and revert work-around for bad initialization detection
The use of zero as a default when getting the current revision, and the failure to gain a lock while reading the current revision could lead to a failure-to-detect-change. This could cause initialization-detection to fail. Change to use INVALID_REVISION and gain lock while attempting to read the current value. 

If this fails to fully fix this issue, then changing the shouldRunInitializationTask() logic to call init() will restore the previous work-around.