fix: re-initialization logic - loadFromXML() is additive
Need to clear the properties before loading since loadFromXML is additive. i.e., it does not clear any properties that are not present in the file. If the load fails, we should logically clear any settings since this is the safest failure mode, so we don't have to attempt to restore to any previously-known values.